INS: do not perform vertical velocity update
The velocity was updated towards zero for indoor mode and
the ublox vertical velocity in outdoor mode. Further testing
has indicated that in indoor mode this overly damps the
velocity estimate. For outdoor mode, it will likely have issues
because of the latency in the GPS.

The baro is a pretty accurate sensor and sufficient to fuse with
the accels to get a good estimate of the vertical velocity.