Quanton: use INS13
The flash allocated to firmware in the STM32F4 in Quanton
is currently not enough to fit PicoC + INS14. This reverts
the INS14 change until the flash is reconfigured.

There is plenty of physical space but it requires a bootloader
update to use it, so this should be done as a separate PR.